+++ tap/plugins/_pytest.py
@@ -18,12 +18,15 @@ def pytest_addoption(parser):
     group.addoption(
         '--tap-stream', default=False, action='store_true', help=_(
             'Stream TAP output instead of the default test runner output.'))
+    group.addoption(
+        '--tap-files', default=False, action='store_true', help=_(
+            'Store all TAP test results into individual files per test case.'))
+    group.addoption(
+        '--tap-combined', default=False, action='store_true', help=_(
+            'Store all TAP test results into a combined output file.'))
     group.addoption('--tap-outdir', metavar='path', help=_(
         'An optional output directory to write TAP files to. '
         'If the directory does not exist, it will be created.'))
-    group.addoption(
-        '--tap-combined', default=False, action='store_true',
-        help=_('Store all TAP test results into a combined output file.'))
 
 
 @pytest.mark.trylast
@@ -55,4 +58,9 @@ def pytest_runtest_logreport(report):
 
 def pytest_unconfigure(config):
     """Dump the results."""
-    tracker.generate_tap_reports()
+    if (
+        config.option.tap_stream or
+        config.option.tap_files or
+        config.option.tap_combined
+    ):
+        tracker.generate_tap_reports()